nix: disable always-allow-substitutes
always-allow-substitutes actually falls back to local build only if it gets 404.

It backfired on boot of offline-booting machine due to home-manager using `nix-env`, and in combination with always-allow-substitutes=true this becomes a requirement of a network connection.

This can be replaced with `--option always-allow-substitutes true` where needed, but only for trusted users.